# docs-standalone/openapi.yaml
openapi: 3.0.3
info:
  title: Examify API Documentation
  version: 1.0.0
  description: >-
    **Examify** is a production-grade, self-hosted AI Exam Generation &
    Assessment Platform built on Next.js 15 (App Router), MariaDB (Prisma ORM),
    Redis, and multi-model AI adapters (Google Gemini, OpenAI, Anthropic Claude,
    and Self-Hosted/vLLM/Ollama).


    ### Core Security & Platform Highlights:

    - **Zero Cloudflare Dependencies**: Self-hosted Redis Sliding Window rate
    limiting and MaxMind GeoIP resolution.

    - **Strict Role-Based Access Control**: `ADMIN` and `TEACHER` roles with
    token blacklisting in Redis.

    - **Two-Factor Authentication (2FA)**: Built-in TOTP (Google Authenticator)
    with QR code setup and challenge verification.

    - **Anti-Prompt Injection Guard**: Advanced multi-tier sanitization and
    jailbreak defense for teacher custom instructions.

    - **Permanent Initialization Lock**: One-time initial setup endpoint locked
    permanently once root admin is created.

    - **Comprehensive Audit Trail**: Immutable system logs recording security
    events, authentication attempts, AI synthesis, and administrative updates.
  contact:
    name: Examify Engineering Team
    email: admin@examify.local
    url: http://localhost:3000
  license:
    name: Proprietary
servers:
  - url: /api/v1
    description: Examify API v1 Base Path
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Provide your JWT bearer token in the Authorization header (e.g.
        `Bearer eyJhbGciOi...`).
  schemas:
...
    StandardError:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Resource not found or validation failure.
        code:
          type: string
          example: VALIDATION_ERROR
        errors:
          type: object
          description: Field-level validation error details
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: d3b07384-d113-460d-9b16-f288e2c4908a
        email:
          type: string
          format: email
          example: teacher@examify.local
        fullName:
          type: string
          example: ThS. Nguyễn Văn An
        role:
          type: string
          enum:
            - ADMIN
            - TEACHER
          example: TEACHER
        school:
          type: string
          example: THPT Chuyên Trần Phú
        twoFactorEnabled:
          type: boolean
          example: false
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
    CognitiveMatrix:
      type: object
      description: Bloom taxonomy percentage distribution (must sum to 100%).
      properties:
        knowledge:
          type: number
          minimum: 0
          maximum: 100
          example: 40
          description: Nhận biết (%)
        comprehension:
          type: number
          minimum: 0
          maximum: 100
          example: 30
          description: Thông hiểu (%)
        application:
          type: number
          minimum: 0
          maximum: 100
          example: 20
          description: Vận dụng (%)
        highApplication:
          type: number
          minimum: 0
          maximum: 100
          example: 10
          description: Vận dụng cao (%)
    QuestionOption:
      type: object
      properties:
        key:
          type: string
          example: A
        text:
          type: string
          example: x = 2 hoặc x = -3
        isCorrect:
          type: boolean
          example: false
    Question:
      type: object
      properties:
        id:
          type: string
          example: q-1
        questionNumber:
          type: integer
          example: 1
        type:
          type: string
          enum:
            - multiple_choice
            - true_false
            - short_answer
            - essay
          example: multiple_choice
        cognitiveLevel:
          type: string
          enum:
            - KNOWLEDGE
            - COMPREHENSION
            - APPLICATION
            - HIGH_APPLICATION
          example: COMPREHENSION
        level:
          type: string
          example: Thông hiểu
        content:
          type: string
          example: "Cho hàm số \\(y = f(x)\\) có bảng biến thiên như hình vẽ. Số điểm cực
            trị của hàm số là:"
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuestionOption"
        correctAnswer:
          type: string
          example: B
        explanation:
          type: string
          example: Đạo hàm đổi dấu 2 lần qua các điểm cực trị.
        points:
          type: number
          example: 0.25
    ExamItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: b1b52817-820d-41d9-adfe-55c8293546c1
        title:
          type: string
          example: Đề thi Khảo sát chất lượng Toán 12 - Học kỳ 1
        subject:
          type: string
          example: Toán học
        grade:
          type: string
          example: Lớp 12
        term:
          type: string
          example: Học kỳ 1
        duration:
          type: integer
          example: 90
        status:
          type: string
          enum:
            - approved
            - review
            - draft
          example: approved
        totalQuestions:
          type: integer
          example: 22
        matrix:
          $ref: "#/components/schemas/CognitiveMatrix"
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /auth/me:
    get:
      tags:
        - Authentication & Profile
      summary: Lấy thông tin tài khoản hiện tại
      description: Trả về hồ sơ cá nhân, vai trò, trường học và trạng thái bảo mật của
        phiên đăng nhập hiện tại.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Thông tin tài khoản xác thực
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user:
                        $ref: "#/components/schemas/UserResponse"
        "401":
          description: Chưa đăng nhập hoặc token đã hết hạn
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
...
  /curriculum:
    get:
      tags:
        - Curriculum
      summary: Lấy cây danh mục chương trình giáo dục K-12
      description: "Trả về toàn bộ phân cấp chương trình giáo dục: Cấp học (THCS,
        THPT) -> Khối lớp (10, 11, 12) -> Môn học -> Học kỳ -> Chủ đề -> Bài
        học."
      parameters:
        - name: level
          in: query
          required: false
          description: "Lọc theo mã cấp học (ví dụ: THPT, THCS)"
          schema:
            type: string
            example: THPT
        - name: grade
          in: query
          required: false
          description: "Lọc theo mã khối lớp (ví dụ: 12, 11, 10)"
          schema:
            type: string
            example: "12"
      responses:
        "200":
          description: Cây danh mục chương trình chuẩn
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                          example: Trung học Phổ thông
                        code:
                          type: string
                          example: THPT
                        grades:
                          type: array
                          items:
                            type: object
  /exams:
    get:
      tags:
        - Exams & Synthesis
      summary: Danh sách đề thi của người dùng (hoặc tất cả nếu là ADMIN)
      description: Lấy danh sách các đề thi đã tạo, trạng thái kiểm duyệt, số câu hỏi
        và ma trận nhận thức tương ứng.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Danh sách đề thi
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ExamItem"
    post:
      tags:
        - Exams & Synthesis
      summary: Tạo hoặc lưu cập nhật đề thi thủ công (Upsert)
      description: Lưu toàn bộ cấu trúc đề thi gồm tiêu đề, môn, khối lớp, danh sách
        câu hỏi, phương án, đáp án và ma trận nhận thức vào cơ sở dữ liệu.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                id:
                  type: string
                  format: uuid
                  description: Để trống nếu tạo mới, hoặc truyền UUID nếu cập nhật đề thi đã có
                title:
                  type: string
                  example: Đề thi Khảo sát Toán 12 Học kỳ 1
                subject:
                  type: string
                  default: Toán học
                  example: Toán học
                grade:
                  type: string
                  default: Lớp 12
                  example: Lớp 12
                term:
                  type: string
                  default: Học kỳ 1
                  example: Học kỳ 1
                duration:
                  type: integer
                  default: 45
                  example: 90
                status:
                  type: string
                  enum:
                    - approved
                    - review
                    - draft
                  default: approved
                matrix:
                  $ref: "#/components/schemas/CognitiveMatrix"
                questions:
                  type: array
                  items:
                    $ref: "#/components/schemas/Question"
      responses:
        "201":
          description: Lưu đề thi thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Đã lưu đề thi thành công vào cơ sở dữ liệu
                  data:
                    type: object
                    properties:
                      examId:
                        type: string
                        format: uuid
...